Cleanup status and workflow generator
