+++ pkg/utils/helpers.go
@@ -1,9 +1,12 @@
 package utils
 
 import (
-	fluxhelmv2beta1 "github.com/fluxcd/helm-controller/api/v2beta1"
-	"sigs.k8s.io/yaml"
+	"fmt"
 	"strings"
+
+	fluxhelmv2beta1 "github.com/fluxcd/helm-controller/api/v2beta1"
+	"helm.sh/helm/v3/pkg/chart"
+	"sigs.k8s.io/yaml"
 )
 
 func ConvertToDNS1123(in string) string {
@@ -39,4 +42,32 @@ func HrToYaml(hr fluxhelmv2beta1.HelmRelease) string {
 	}
 
 	return string(b)
+}
+
+func TemplateContainsYaml(ch *chart.Chart) (bool, error) {
+	if ch == nil {
+		return false, fmt.Errorf("chart cannot be nil")
+	}
+
+	for _, f := range ch.Templates {
+		if IsFileYaml(f.Name) {
+			return true, nil
+		}
+	}
+	return false, nil
+}
+
+func IsFileYaml(f string) bool {
+	f = strings.ToLower(f)
+	if strings.HasSuffix(f, "yml") || strings.HasSuffix(f, "yaml") {
+		return true
+	}
+	return false
+}
+
+func AddAppChartNameToFile(name, a string) string {
+	prefix := "templates/"
+	name = strings.TrimPrefix(name, prefix)
+	name = a + "_" + name
+	return prefix + name
 }
